+++ parent/document/net.bluemind.document.sqlschema/sql/document-schema.1.0.0.sql
@@ -5,7 +5,7 @@ create table t_document (
   name varchar(256),
   description text,
   mime varchar(64),
-  item_id int4 references t_container_item(id) on delete cascade
+  item_id bigint references t_container_item(id) on delete cascade
 );
