+++ parent/group/net.bluemind.group.sqlschema/sql/group-schema.1.0.0.sql
@@ -1,7 +1,7 @@
 CREATE EXTENSION IF NOT EXISTS hstore WITH schema pg_catalog;
 
 CREATE TABLE t_group (
-  item_id       int4 references t_container_item(id) PRIMARY KEY,
+  item_id       bigint references t_container_item(id) PRIMARY KEY,
   container_id  int4 references t_container(id),
 
   name          text not null,
@@ -18,27 +18,27 @@ CREATE TABLE t_group (
 );
 
 CREATE TABLE t_group_groupmember (
-  group_parent_id int4 references t_group(item_id),
-  group_child_id int4 references t_group(item_id),
+  group_parent_id bigint references t_group(item_id),
+  group_child_id bigint references t_group(item_id),
 
   PRIMARY KEY(group_parent_id, group_child_id)
 );
 
 CREATE TABLE t_group_usermember  (
-  group_id     int4 references t_group(item_id),
-  user_id     int4 references t_container_item(id) on delete cascade,
+  group_id     bigint references t_group(item_id),
+  user_id     bigint references t_container_item(id) on delete cascade,
   PRIMARY KEY(group_id, user_id)
 );
 
 CREATE TABLE t_group_externalusermember  (
-  group_id     int4 references t_group(item_id),
-  external_user_id     int4 references t_container_item(id) on delete cascade,
+  group_id     bigint references t_group(item_id),
+  external_user_id     bigint references t_container_item(id) on delete cascade,
   PRIMARY KEY(group_id, external_user_id)
 );
 
 CREATE TABLE t_group_flat_members (
-  group_id        int4 references t_group(item_id),
-  user_id     int4 references t_container_item(id) on delete cascade,
+  group_id        bigint references t_group(item_id),
+  user_id     bigint references t_container_item(id) on delete cascade,
   PRIMARY KEY(group_id, user_id)
 );
